249473 - UPD cleanup to not run on localhost

+++ lib/hix/exe/new.rb
@@ -96,6 +96,8 @@ module Hix
       end
 
       def cleanup
+        return if Hix::CONFIG["env"] == "local"
+
         FileUtils.rm_rf(version_cache_path)
       end
